requirements.txt
streamlit==1.32.2
pillow==10.2.0
openai==1.63.2
pydantic==2.11.0
mcp==1.26.0
python-dotenv==1.1.0
anyio==4.3.0
httpx==0.27.2
